cast values
Former-commit-id: 5c1cfda328866df637cad6d315a391e8f3bc7075

+++ src/VisIt/libsim/visit_libsim.cc
@@ -423,7 +423,7 @@ void visit_InitLibSim( visit_simulation_data *sim )
     sim->nodeIndex = 0;
 
     sim->maxNodes = 1;
-    sim->maxCores = std::max( sim->myworld->nRanks(),
+    sim->maxCores = std::max( (unsigned int) sim->myworld->nRanks(),
 			      sysGetNumSockets() * sysGetNumCoresPerSockets() );
 
     sim->nodeStart.push_back( 0 );
